moved entities to Shared Library
