+++ src/Claimini.Api/Controllers/TestController.cs
@@ -1,8 +1,7 @@
 using System;
 using System.Collections.Generic;
-using Claimini.Api.Data;
-using Claimini.Api.Data.Dto;
 using Claimini.Api.Services;
+using Claimini.Shared;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Claimini.Api.Controllers
