+++ src/Claimini.Api/Data/IMongoDbContext.cs
@@ -1,4 +1,5 @@
-﻿using MongoDB.Driver;
+using Claimini.Shared;
+using MongoDB.Driver;
 
 namespace Claimini.Api.Data
 {
