+++ src/Claimini.Api/Repository/IMongoRepository.cs
@@ -8,6 +8,7 @@ using System.Collections.Generic;
 using System.Linq.Expressions;
 using System.Threading.Tasks;
 using Claimini.Api.Data;
+using Claimini.Shared;
 
 namespace Claimini.Api.Repository
 {
