+++ src/Claimini.Shared/Invoice.cs
@@ -7,13 +7,11 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
-using Claimini.Shared;
-using Microsoft.EntityFrameworkCore.Scaffolding.Metadata;
 using MongoDB.Bson;
 using MongoDB.Bson.Serialization.Attributes;
 using Newtonsoft.Json;
 
-namespace Claimini.Api.Data
+namespace Claimini.Shared
 {
     /// <summary>
     /// Represents an Invoice, which contains <see cref="InvoiceItem"/>s
